requirements.txt
contextily==1.1.0
sgp4==2.20
matplotlib==3.3.2
nonebot==1.8.0
Shapely==1.7.1
geopandas==0.9.0
pypinyin==0.42.0
Pillow==8.3.1
